[FEATURE] introduce languageAspect and pages.l18n_cfg information to settings
In order to cope for different l10n stuff it is important to sum up all info.
This allows to get a pseudo languageAspect which is the core way to know the
correct fallback for the currently selected language. As we focus on single
page additional pages.l18n_cfg is added which can change the language mode
per page. As this can create a lot of combinations, this introduction of an
adapted languageAspect could work as single point of complexity.

This is to support functions for language views (e.g. only show translation
with id 2 and not all) and could be either used
* while doing the output in fluid to simply filter the view, or
* while actually fetching localizations and localizationActions thus adding
  to the sDEF/vDEF complexity (which would be possible, but probably u
  nnecessary)

It uses the 9LTS introduced Context/LanguageAspect architecture, but provides
sane defaults for 8LTS (i.e. the recommendations of core mixed with the actual
current implementation of EXT:templavoilaplus).