[FEATURE] add TSconfig additionalRecordData.<table> for BE node data attributes
It would be a benefit if like in TVP7 column values of nodes could have an inpact in BE preview rendering. In TVP7 this was hardcoded to `layout` I propose to make it TSconfig configurable, e.g. use

`mod.web_txtemplavoilaplusLayout.additionalRecordData.tt_content = layout,imageorient`

to add data-layout and data-imageorient to a tt_content node in order to allow further styling or custom JavaScript